+++ pyController/.env-example
@@ -4,7 +4,7 @@
 FACTORY_IP='0.0.0.0'
 # modbus port. Default port is 502
 FACTORY_PORT=502
-# Factory simulation True/False. Uses FactorySim2 instead of FactoryModbus
+# Factory simulation True/False. Uses FactorySim2 instead of Factory
 FACTORY_SIM=True
 
 ################
