+++ pyController/factory/factory_sim2.py
@@ -1,6 +1,6 @@
 
 """This simulates the input and output of the real factory module
-   Used as a drop in replacement for factoryModbus when the PLC is unavailable"""
+   Used as a drop in replacement for the Factory module when the PLC is unavailable"""
 
 import logging
 import threading
@@ -12,7 +12,7 @@ logger.setLevel(logging.DEBUG) # sets default logging level for module
 
 class FactorySim2():
     """This class simulates the input and output of the real factory class
-       Used as a drop in replacement for factoryModbus when the PLC is unavailable"""
+       Used as a drop in replacement for Factory class when the PLC is unavailable"""
 
     def __init__(self):
         """ Initialize class instance"""
